umgebung/testumgebung/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
passlib==1.7.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0
waitress==2.1.2
Werkzeug==2.2.2
XlsxWriter==3.0.8
